build_automation/requirements/base.txt
django-environ==0.4.5
djangorestframework==3.9.0
psycopg2-binary==2.7.5
tox
celery==4.2.1
django-filter